requirements.txt
asgiref==3.2.10
Django==3.1.1
django-environ==0.4.5
django-filter==2.3.0
djangorestframework==3.11.1
install==1.3.3
model-mommy==2.0.0
pkg-resources==0.0.0
psycopg2-binary==2.8.6
pytz==2020.1
sqlparse==0.3.1
